# docs/openapi.yml
openapi: 3.1.0
info:
  title: Aegiscan API
  summary: Aegiscan API
  description: Aegiscan is the open source Tines / Splunk SOAR alternative. You can
    operate Aegiscan in headless mode by using the API to create, manage, and run
    workflows.
  termsOfService: https://docs.google.com/document/d/e/2PACX-1vQvDe3SoVAPoQc51MgfGCP71IqFYX_rMVEde8zC4qmBCec5f8PLKQRdxa6tsUABT8gWAR9J-EVs2CrQ/pub
  contact:
    name: Aegiscan Founders
    email: founders@digi-trans.org
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  version: '1'
paths:
...
  /workflows/{workflow_id}/tags:
    get:
      tags:
      - workflows
      summary: List Tags
      description: List all tags for a workflow.
      operationId: workflows-list_tags
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRead'
                title: Response Workflows-List Tags
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - workflows
      summary: Add Tag
      description: Add a tag to a workflow.
      operationId: workflows-add_tag
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowTagCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workflows/{workflow_id}/tags/{tag_id}:
    delete:
      tags:
      - workflows
      summary: Remove Tag
      operationId: workflows-remove_tag
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: tag_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: Tag Id
      - name: workflow_id
        in: path
        required: true
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /tags:
    get:
      tags:
      - tags
      summary: List Tags
      description: List all tags for the current workspace.
      operationId: tags-list_tags
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRead'
                title: Response Tags-List Tags
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - tags
      summary: Create Tag
      description: Create a new tag.
      operationId: tags-create_tag
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tags/{tag_id}:
    get:
      tags:
      - tags
      summary: Get Tag
      description: Get a specific tag by ID.
      operationId: tags-get_tag
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: tag_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: Tag Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - tags
      summary: Update Tag
      description: Update an existing tag.
      operationId: tags-update_tag
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: tag_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: Tag Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - tags
      summary: Delete Tag
      description: Delete a tag.
      operationId: tags-delete_tag
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: tag_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: Tag Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TagCreate:
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 1
          title: Name
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
          description: Hex color code
      type: object
      required:
      - name
      title: TagCreate
      description: Model for creating new tags with validation.
    TagRead:
      properties:
        id:
          type: string
          format: uuid4
          title: Id
        name:
          type: string
          maxLength: 50
          minLength: 1
          title: Name
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
          description: Hex color code
      type: object
      required:
      - id
      - name
      title: TagRead
      description: Model for reading tag data with validation.
    TagUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 50
            minLength: 1
          - type: 'null'
          title: Name
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
          description: Hex color code
      type: object
      title: TagUpdate
      description: Model for updating existing tags with validation.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WorkflowTagCreate:
      properties:
        tag_id:
          type: string
          format: uuid4
          title: Tag Id
      type: object
      required:
      - tag_id
      title: WorkflowTagCreate
...
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: fastapiusersauth
    APIKeyHeader:
      type: apiKey
      in: header
      name: x-aegiscan-service-key
...